big big lobotomy big lobotomy

+++ src/main/java/com/hbm/items/armor/JetpackBreak.java
@@ -32,11 +32,11 @@ public class JetpackBreak extends JetpackBase {
 	}
 
 	public void onArmorTick(World world, EntityPlayer player, ItemStack stack) {
-		
+
 		HbmPlayerProps props = HbmPlayerProps.getData(player);
-		
+
 		if(!world.isRemote) {
-			
+
 			if(getFuel(stack) > 0 && (props.isJetpackActive() || (!player.onGround && !player.isSneaking() && props.enableBackpack))) {
 
 	    		NBTTagCompound data = new NBTTagCompound();
@@ -48,20 +48,20 @@ public class JetpackBreak extends JetpackBase {
 
 		if(getFuel(stack) > 0) {
 			float gravity = Math.max(CelestialBody.getBody(world).getSurfaceGravity(), AstronomyUtil.STANDARD_GRAVITY);
-			float thrustMultiplier = gravity * AstronomyUtil.PLAYER_GRAVITY_MODIFIER;
-			
+			float thrustMultiplier = (float) (gravity * AstronomyUtil.PLAYER_GRAVITY_MODIFIER);
+
 			if(props.isJetpackActive()) {
 				player.fallDistance = 0;
-				
+
 				if(player.motionY < 0.4D)
 					player.motionY += 0.1D * thrustMultiplier;
-				
+
 				world.playSoundEffect(player.posX, player.posY, player.posZ, "hbm:weapon.flamethrowerShoot", 0.25F, 1.5F);
 				this.useUpFuel(player, stack, 5);
-				
+
 			} else if(!player.isSneaking() && !player.onGround && props.enableBackpack) {
 				player.fallDistance = 0;
-				
+
 				if(player.motionY < -1 * thrustMultiplier)
 					player.motionY += 0.2D * thrustMultiplier;
 				else if(player.motionY < -0.1 * thrustMultiplier)
@@ -71,20 +71,20 @@ public class JetpackBreak extends JetpackBase {
 
 				player.motionX *= 1.025D;
 				player.motionZ *= 1.025D;
-				
+
 				world.playSoundEffect(player.posX, player.posY, player.posZ, "hbm:weapon.flamethrowerShoot", 0.25F, 1.5F);
 				this.useUpFuel(player, stack, 10);
 			}
 		}
     }
-    
+
     @SideOnly(Side.CLIENT)
     public void addInformation(ItemStack stack, EntityPlayer player, List list, boolean ext) {
 
     	list.add("Regular jetpack that will automatically hover mid-air.");
     	list.add("Sneaking will stop hover mode.");
     	list.add("Hover mode will consume less fuel and increase air-mobility.");
-    	
+
     	super.addInformation(stack, player, list, ext);
     }
 }
